+++ src/library/DIPS.Mobile.UI/API/Library/Android/DUI.cs
@@ -1,16 +1,10 @@
-using System.Diagnostics;
-using Android.App;
-using Android.Content;
-using Android.Content.PM;
 using Android.Graphics.Drawables;
-using Android.Hardware.Display;
 using Android.Views;
 using AndroidX.Core.SplashScreen;
 using DIPS.Mobile.UI.API.Camera.Shared.Android;
-using DIPS.Mobile.UI.Components.Pickers.DatePicker.Android;
+using DIPS.Mobile.UI.API.Library.Android;
 using DIPS.Mobile.UI.Components.Pickers.DatePicker.Service;
-using DIPS.Mobile.UI.Components.Pickers.TimePicker.Android;
-using Microsoft.Maui.Controls.PlatformConfiguration;
+using Microsoft.Maui.Handlers;
 using Microsoft.Maui.Platform;
 using TimePickerService = DIPS.Mobile.UI.Components.Pickers.TimePicker.TimePickerService;
 
@@ -19,7 +13,7 @@ namespace DIPS.Mobile.UI.API.Library;
 
 public static partial class DUI
 {
-    public static void Init(Android.App.Activity activity)
+    public static void Init(global::Android.App.Activity activity)
     {
         SplashScreen.InstallSplashScreen(activity);
         
@@ -27,6 +21,15 @@ public static partial class DUI
         // MAUI has a bug after rewriting modals to use DialogFragment
         // Workaround found here: https://github.com/CommunityToolkit/Maui/issues/2370#issuecomment-2552701081
         activity.GetFragmentManager()?.RegisterFragmentLifecycleCallbacks(new FragmentLifeCycleCallback(), false);
+        
+        // Only add the handlers if the MainActivity is derived from our MauiAppCompatActivity
+        if (activity is MauiAppCompatActivity)
+        {
+            PageHandler.Mapper.ReplaceMapping<ContentPage, IPageHandler>(nameof(ContentPage.HideSoftInputOnTapped), HideSoftInputOnTapHandlerMappings.MapHideSoftInputOnTapped);
+            EntryHandler.Mapper.ReplaceMapping<Entry, IEntryHandler>(nameof(VisualElement.IsFocused), HideSoftInputOnTapHandlerMappings.MapInputIsFocused);
+            EditorHandler.Mapper.ReplaceMapping<Editor, IEditorHandler>(nameof(VisualElement.IsFocused), HideSoftInputOnTapHandlerMappings.MapInputIsFocused);
+            SearchBarHandler.Mapper.ReplaceMapping<SearchBar, ISearchBarHandler>(nameof(VisualElement.IsFocused), HideSoftInputOnTapHandlerMappings.MapInputIsFocused);
+        }
     }
     
     private static partial void PlatformInit()